requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
pandas==2.2.2
openpyxl==3.1.5
xlrd==2.0.1 # for reading legacy .xls
requests==2.32.3